add game logic & more optimized
picks aren't placed right (they're in the same spot as the slot they're for) and pick selection needs a rework to support more than 4 picks, and not overlap the outer ring. only Easy is playable, but it is completable on Easy. Also added icon, app info, and a modified applist for it to show up in the chronomark's menu.